% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COA.R
\name{coa_GetBodyDirection}
\alias{coa_GetBodyDirection}
\title{获取COA表体走向}
\usage{
coa_GetBodyDirection(
  erpToken = "C0426D23-1927-4314-8736-A74B2EF7A039",
  FTemplateNumber = "M021"
)
}
\arguments{
\item{erpToken}{ERP口令}

\item{FTemplateNumber}{单据编号}
}
\value{
返回值
}
\description{
获取COA表体走向
}
\examples{
coa_GetBodyDirection()
}
