% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COA.R
\name{coa_GetCustomerName}
\alias{coa_GetCustomerName}
\title{获取客户名称}
\usage{
coa_GetCustomerName(
  erpToken = "C0426D23-1927-4314-8736-A74B2EF7A039",
  FBillNo = "XSCKD-102-20250521-0004"
)
}
\arguments{
\item{erpToken}{ERP口令}

\item{FBillNo}{单据编号}
}
\value{
返回值
}
\description{
获取客户名称
}
\examples{
coa_GetCustomerName()
}
