% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COA.R
\name{coa_GetFDate}
\alias{coa_GetFDate}
\title{获取日期}
\usage{
coa_GetFDate(
  erpToken = "C0426D23-1927-4314-8736-A74B2EF7A039",
  FBillNo = "XSCKD-102-20250521-0004"
)
}
\arguments{
\item{erpToken}{ERP口令}

\item{FBillNo}{单据编号}
}
\value{
返回值
}
\description{
获取日期
}
\examples{
coa_GetFDate()
}
