% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COA.R
\name{coa_GetTemplateNumber}
\alias{coa_GetTemplateNumber}
\title{通过销售出现单获取COA模板号}
\usage{
coa_GetTemplateNumber(
  erpToken = "C0426D23-1927-4314-8736-A74B2EF7A039",
  FBillNo = "XSCKD-102-20250521-0004"
)
}
\arguments{
\item{erpToken}{ERP口令}

\item{FBillNo}{单据编号}
}
\value{
返回值
}
\description{
通过销售出现单获取COA模板号
}
\examples{
coa_GetTemplateNumber()
}
