% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COA.R
\name{coa_meta}
\alias{coa_meta}
\title{返回元数据信息}
\usage{
coa_meta(
  erpToken = "C0426D23-1927-4314-8736-A74B2EF7A039",
  FTemplateNumber = "M001_100_IMP",
  FCOATableType = "billHead"
)
}
\arguments{
\item{erpToken}{ERP口令}

\item{FTemplateNumber}{模板号}

\item{FCOATableType}{数据类型}
}
\value{
返回值
}
\description{
返回元数据信息
}
\examples{
coa_meta()
}
