% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COA.R
\name{coa_pdf}
\alias{coa_pdf}
\title{将销售出库单数据写入EXCEL并上传到OSS对象存储}
\usage{
coa_pdf(
  erpToken = "C0426D23-1927-4314-8736-A74B2EF7A039",
  FBillNo = "XSCKD-100-20250523-0001",
  outputDir = getwd(),
  delete_localFiles = 0
)
}
\arguments{
\item{erpToken}{ERP口令}

\item{FBillNo}{单据编号}

\item{outputDir}{输出地址}

\item{delete_localFiles}{是否删除本地文件}
}
\value{
返回值
}
\description{
将销售出库单数据写入EXCEL并上传到OSS对象存储
}
\examples{
coa_pdf()
}
