% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COA.R
\name{coa_SyncAll}
\alias{coa_SyncAll}
\title{获取单据清单}
\usage{
coa_SyncAll(
  erpToken = "C0426D23-1927-4314-8736-A74B2EF7A039",
  outputDir = getwd(),
  delete_localFiles = 0
)
}
\arguments{
\item{erpToken}{ERP 口令}
}
\value{
返回值
}
\description{
获取单据清单
}
\examples{
coa_TaskToSync()
}
