% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomerMaterialCorrespondence.R
\name{CustomerMaterialCorrespondence_delete}
\alias{CustomerMaterialCorrespondence_delete}
\title{删除覆盖前删除}
\usage{
CustomerMaterialCorrespondence_delete(erp_token)
}
\value{
无返回值
}
\description{
删除覆盖前删除
}
\examples{
CustomerMaterialCorrespondence_delete()
}
