% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomerMaterialCorrespondence.R
\name{CustomerMaterialCorrespondence_inputdelete}
\alias{CustomerMaterialCorrespondence_inputdelete}
\title{删除客户为空}
\usage{
CustomerMaterialCorrespondence_inputdelete(erp_token)
}
\value{
无返回值
}
\description{
删除客户为空
}
\examples{
CustomerMaterialCorrespondence_inputdelete()
}
