% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomerMaterialCorrespondence.R
\name{CustomerMaterialCorrespondence_insert}
\alias{CustomerMaterialCorrespondence_insert}
\title{插入正式表}
\usage{
CustomerMaterialCorrespondence_insert(erp_token)
}
\value{
无返回值
}
\description{
插入正式表
}
\examples{
CustomerMaterialCorrespondence_insert()
}
