% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomerMaterialCorrespondence.R
\name{CustomerMaterialCorrespondence_select}
\alias{CustomerMaterialCorrespondence_select}
\title{查询客户物料对应}
\usage{
CustomerMaterialCorrespondence_select(erp_token)
}
\value{
无返回值
}
\description{
查询客户物料对应
}
\examples{
CustomerMaterialCorrespondence_select()
}
