% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomerMaterialCorrespondence.R
\name{CustomerMaterialCorrespondence_truncate}
\alias{CustomerMaterialCorrespondence_truncate}
\title{清空客户物料对应}
\usage{
CustomerMaterialCorrespondence_truncate(erp_token)
}
\value{
无返回值
}
\description{
清空客户物料对应
}
\examples{
CustomerMaterialCorrespondence_truncate()
}
