% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomerTemplates.R
\name{CustomerTemplates_delete}
\alias{CustomerTemplates_delete}
\title{覆盖前删除}
\usage{
CustomerTemplates_delete(erp_token)
}
\value{
无返回值
}
\description{
覆盖前删除
}
\examples{
CustomerTemplates_delete()
}
