% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomerTemplates.R
\name{CustomerTemplates_insert}
\alias{CustomerTemplates_insert}
\title{删除客户对应模板表为空}
\usage{
CustomerTemplates_insert(erp_token)
}
\value{
无返回值
}
\description{
删除客户对应模板表为空
}
\examples{
CustomerTemplates_insert()
}
