% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomerTemplates.R
\name{CustomerTemplates_truncate}
\alias{CustomerTemplates_truncate}
\title{清空客户对应模板表}
\usage{
CustomerTemplates_truncate(erp_token)
}
\value{
无返回值
}
\description{
清空客户对应模板表
}
\examples{
CustomerTemplates_truncate()
}
